TIdHTTPRangeStream, TIdHTTPResponseInfo, and TIdCustomHTTPServer now use IdDisposeAndNil() instead of FreeAndNil() if a TStream is marked as owned and needs to be freed.
If the client sends an 'Authorization' request header, but it is not processed, TIdHTTPServer now sends an HTTP 401 response instead of closing the socket connection.

TIdHTTPServer now fires the OnDoneWithPostStream event after sending a response, if the ARequestInfo.PostStream property is not nil.

When sending a response to a client, TIdHTTPResponseInfo.SetHeaders() now sets 'WWW-Authenticate' and 'Proxy-Authenticate' headers if the TIdHTTPResponseInfo.WWWAuthenticate and/or TIdHTTPResponseInfo.ProxyAuthenticate properties are not empty.

git-svn-id: https://svn.atozed.com:444/svn/Indy10/trunk@5446 421b718a-868e-4740-9b41-4026df17543c